Separated domain model from request transfer objects

+++ Eetfestijnkassasystem.Api/Controllers/OrdersController.cs
@@ -5,6 +5,8 @@ using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
 using Eetfestijnkassasystem.Api.Data;
 using Eetfestijnkassasystem.Shared.Model;
+using Eetfestijnkassasystem.Shared.DTO;
+using System;
 
 namespace Eetfestijnkassasystem.Api.Controllers
 {
@@ -23,21 +25,30 @@ namespace Eetfestijnkassasystem.Api.Controllers
         [HttpGet]
         public async Task<ActionResult<IEnumerable<Order>>> GetOrder()
         {
-            return await _context.Order.ToListAsync();
+            List<OrderModel> orders = await _context.Order
+                                                    .Include(o => o.OrderMenuItems)
+                                                    .ThenInclude(o => o.MenuItem)
+                                                    .ToListAsync();
+
+            return orders.Select(o => o.ToTransferObject()).ToList();
         }
 
         // GET: api/Orders/5
         [HttpGet("{id}")]
         public async Task<ActionResult<Order>> GetOrder(int id)
         {
-            var order = await _context.Order.FindAsync(id);
+            if (id == 0)
+                return BadRequest("Order id not specified");
 
-            if (order == null)
-            {
-                return NotFound();
-            }
+            OrderModel model = await _context.Order
+                                             .Include(o => o.OrderMenuItems)
+                                             .ThenInclude(o => o.MenuItem)
+                                             .FirstOrDefaultAsync(o => o.Id == id);
 
-            return order;
+            if (model == null)
+                return NotFound($"Order with id {id} not found");
+
+            return model.ToTransferObject();
         }
 
         // PUT: api/Orders/5
@@ -46,15 +57,26 @@ namespace Eetfestijnkassasystem.Api.Controllers
         [HttpPut("{id}")]
         public async Task<IActionResult> PutOrder(int id, Order order)
         {
-            if (id != order.Id)
-            {
-                return BadRequest();
-            }
-
-            _context.Entry(order).State = EntityState.Modified;
-
             try
             {
+                if (id == 0 || id != order.Id)
+                    return BadRequest();
+
+                OrderModel model = await _context.Order.FindAsync(id);
+
+                if (model == null)
+                    return NotFound();
+
+                model.CustomerName = order.CustomerName;
+                model.Comment = order.Comment;
+                model.Seating = order.Seating;
+                model.Payment = order?.Payment.ToModelEntity();
+
+                model.OrderMenuItems.Clear();
+                model.OrderMenuItems.AddRange(await CreateOrderMenuItemsAsync(order.MenuItems));
+                
+                _context.Entry(model).State = EntityState.Modified;
+
                 await _context.SaveChangesAsync();
             }
             catch (DbUpdateConcurrencyException)
@@ -78,40 +100,109 @@ namespace Eetfestijnkassasystem.Api.Controllers
         [HttpPost]
         public async Task<ActionResult<Order>> PostOrder(Order order)
         {
-            for (int i = 0; i < order.OrderMenuItems.Count; i++)
+            try
             {
-                var existingMenuItem = await _context.MenuItem.FirstOrDefaultAsync(
-                    o => o.Name.ToLower().Equals(order.OrderMenuItems[i].MenuItem.Name.ToLower()));
+                OrderModel newOrderModel = order.ToModelEntity();
 
-                if (existingMenuItem != null)
-                    order.OrderMenuItems[i].MenuItem = existingMenuItem;
+                //List<OrderMenuItem> newOrderMenuItems = new List<OrderMenuItem>();
+
+                //foreach (MenuItem mito in order.MenuItems)
+                //{
+                //    MenuItemModel mim = await _context.MenuItem.FindAsync(mito.Id);
+
+                //    if (mim == null)
+                //        mim = await _context.MenuItem.FirstOrDefaultAsync(
+                //            o => o.Name.Replace(" ", "").ToLower().Equals(mito.Name.Replace(" ", "").ToLower()));
+
+                //    if (mim == null)
+                //        mim = mito.ToModelEntity();
+
+                //    if (newOrderMenuItems.Any(o => o.MenuItem == mim))
+                //        newOrderMenuItems.FirstOrDefault(o => o.MenuItem == mim).MenuItemCount++;
+                //    else
+                //        newOrderMenuItems.Add(new OrderMenuItem() { MenuItem = mim, MenuItemCount = 1 });
+                //}
+
+                List<OrderMenuItem> newOrderMenuItems = await CreateOrderMenuItemsAsync(order.MenuItems);
+
+                newOrderModel.OrderMenuItems.AddRange(newOrderMenuItems);
+
+                _context.Order.Add(newOrderModel);
+                await _context.SaveChangesAsync();
+
+                Order createdOrder = newOrderModel.ToTransferObject();
+
+                return CreatedAtAction("GetOrder", new { id = newOrderModel.Id }, createdOrder);
             }
-
-            _context.Order.Add(order);
-            await _context.SaveChangesAsync();
-
-            return CreatedAtAction("GetOrder", new { id = order.Id }, order);
+            catch (Exception ex)
+            {
+                throw;
+            }
         }
 
         // DELETE: api/Orders/5
         [HttpDelete("{id}")]
         public async Task<ActionResult<Order>> DeleteOrder(int id)
         {
+            if (id == 0)
+                return BadRequest("Order id not specified");
+
             var order = await _context.Order.FindAsync(id);
+
             if (order == null)
-            {
-                return NotFound();
-            }
+                return NotFound($"Order with id {id} not found");
 
             _context.Order.Remove(order);
             await _context.SaveChangesAsync();
 
-            return order;
+            return order.ToTransferObject(); // return void?
         }
 
         private bool OrderExists(int id)
         {
             return _context.Order.Any(e => e.Id == id);
+        }
+
+        private async Task<List<OrderMenuItem>> CreateOrderMenuItemsAsync(List<MenuItem> menuItems)
+        {
+            List<OrderMenuItem> orderMenuItems = new List<OrderMenuItem>();
+
+            foreach (MenuItem menuItem in menuItems)
+            {
+                //MenuItemModel model = await _context.MenuItem.FindAsync(menuItem.Id);
+
+                //if (model == null)
+                //{ 
+                //    model = await _context.MenuItem.FirstOrDefaultAsync(
+                //        o => o.Name.Replace(" ", "").ToLower().Equals(menuItem.Name.Replace(" ", "").ToLower()));
+                //}
+
+                //if (model == null)
+                //    model = menuItem.ToModelEntity();
+
+                MenuItemModel model = await FetchOrCreateMenuItemModelAsync(menuItem);
+
+                if (orderMenuItems.Any(o => o.MenuItem == model))
+                    orderMenuItems.FirstOrDefault(o => o.MenuItem == model).MenuItemCount++;
+                else
+                    orderMenuItems.Add(new OrderMenuItem() { MenuItem = model, MenuItemCount = 1 });
+            }
+
+            return orderMenuItems;
+        }
+
+        private async Task<MenuItemModel> FetchOrCreateMenuItemModelAsync(MenuItem menuItem)
+        {
+            MenuItemModel model = await _context.MenuItem.FindAsync(menuItem.Id);
+
+            if (model == null)
+                model = await _context.MenuItem.FirstOrDefaultAsync(o => o.Name.Replace(" ", "").ToLower()
+                                                                          .Equals(menuItem.Name.Replace(" ", "").ToLower()));
+
+            if (model == null)
+                model = menuItem.ToModelEntity();
+
+            return model;
         }
     }
 }
